refator(Member): add Phone class
