+++ src/main/java/com/employee/app/view/InputView.java
@@ -8,7 +8,6 @@ public class InputView {
     private static final String MEMBER_NAME_QUESTION = "사용자 이름을 입력하세요";
     private static Scanner scanner = new Scanner(System.in);
 
-
     public static Name askMemberName() {
         System.out.println(MEMBER_NAME_QUESTION);
         return new Name(scanner.nextLine());
